# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dead Man's Slot Machine API
  description: Wild West themed slot machine game API (simulation only, no real money)
  version: 0.5.0
  contact:
    name: Tin Minarik
    email: tinminarik00@gmail.com
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
paths:
...
  /game/list:
    get:
      tags: [Game]
      summary: List available slot games
      description: Returns all slot game configs including symbols, ways, bet options, and feature flags.
      operationId: getGameList
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of available games
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: dead-mans-gun
                        name:
                          type: string
                          example: "Dead Man's Gun"
                        description:
                          type: string
                        theme:
                          type: string
                          enum: [gunslinger, treasure, desert, canyon]
                        reels:
                          type: integer
                          example: 5
                        rows:
                          type: integer
                          example: 4
                        ways:
                          type: integer
                          example: 1024
                        betOptions:
                          type: array
                          items:
                            type: number
                          example: [10, 20, 50, 100, 200, 500]
                        symbols:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              emoji:
                                type: string
                              name:
                                type: string
                              type:
                                type: string
                                enum: [wild, scatter, regular, bonus]
                              payouts:
                                type: object
                                nullable: true
                              scatterPayouts:
                                type: object
                                nullable: true
  /game/{id}:
    get:
      tags: [Game]
      summary: Get a single game config
      operationId: getGame
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: dead-mans-gun
      responses:
        "200":
          description: Game config
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    type: object
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /game/spin:
    post:
      tags: [Game]
      summary: Spin the slot machine
      description: |
        Deducts bet from CHIPS wallet, generates a random result, credits any winnings.
        Supports auto-spin with `spins` parameter (1, 5, 10, or 100).
        Single spin (spins=1) returns a flat result. Multi-spin returns an array of results plus a summary.
      operationId: spin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gameId, betAmount]
              properties:
                gameId:
                  type: string
                  example: dead-mans-gun
                betAmount:
                  type: number
                  description: Must be one of the game's betOptions
                  example: 20
                spins:
                  type: integer
                  description: "Number of spins to perform (auto-spin). Default: 1"
                  enum: [1, 5, 10, 100]
                  default: 1
      responses:
        "200":
          description: Spin result
          content:
            application/json:
              schema:
                type: object
                properties:
                  gameId:
                    type: string
                  grid:
                    type: array
                    description: "grid[reel][row] = symbolId"
                    items:
                      type: array
                      items:
                        type: string
                  wins:
                    type: array
                    items:
                      type: object
                      properties:
                        symbolId:
                          type: string
                        count:
                          type: integer
                        ways:
                          type: integer
                        positions:
                          type: array
                          items:
                            type: array
                            items:
                              type: integer
                        payout:
                          type: number
                        isScatter:
                          type: boolean
                  totalWin:
                    type: number
                  totalBet:
                    type: number
                  balance:
                    type: number
                    description: CHIPS balance after spin
                  featureTriggered:
                    type: object
                    nullable: true
                    description: "Feature trigger info (null if none). Reserved for future mini-games."
                  symbolMap:
                    type: object
                    description: "Map of symbolId → {emoji, name, type}"
                  xpEarned:
                    type: integer
                    description: XP earned from this spin
                  levelUp:
                    type: object
                    nullable: true
                    description: "Level-up info (null if no level-up)"
                    properties:
                      newLevel:
                        type: integer
                      reward:
                        type: integer
                        description: CHIPS rewarded for leveling up
        "200  (multi-spin)":
          description: "Multi-spin result (when spins > 1)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  spins:
                    type: array
                    description: Individual spin results
                    items:
                      type: object
                      properties:
                        grid:
                          type: array
                          items:
                            type: array
                            items:
                              type: string
                        wins:
                          type: array
                          items:
                            type: object
                        totalWin:
                          type: number
                        totalBet:
                          type: number
                        symbolMap:
                          type: object
                  summary:
                    type: object
                    properties:
                      count:
                        type: integer
                        description: Number of spins performed
                      totalBet:
                        type: number
                      totalWin:
                        type: number
                      netResult:
                        type: number
                        description: totalWin - totalBet
                      wins:
                        type: integer
                        description: Number of winning spins
                      biggestWin:
                        type: number
                      xpEarned:
                        type: integer
                      levelUp:
                        type: object
                        nullable: true
                        properties:
                          newLevel:
                            type: integer
                          reward:
                            type: integer
                  balance:
                    type: number
                    description: CHIPS balance after all spins
        "400":
          description: Invalid bet or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Game not found
...